// web/lib/api/generated/model/bankDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface BankDetails {
  /** @maxLength 200 */
  bank_account_name: string;
  readonly bank_account_number_masked: string;
  /** @maxLength 11 */
  bank_ifsc: string;
  /** @maxLength 120 */
  bank_name: string;
  /** @maxLength 120 */
  bank_branch: string;
  /** @maxLength 320 */
  upi_id: string;
}